+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 *** 0.39 *** 
 - Sensors now are "state_class":"measurement"
 - No need to config MQTT user and password.
+- Estimated Power Use (BETA)
 
 *** 0.38 *** 
 - fix returned a non-zero code on install